Move the table and list data type out of the data folder
