+++ src/components/GameArea/type.ts
@@ -1,6 +1,6 @@
 import { GameDataType } from "../../__generated__/graphql";
-import ListAPIData from "../../types/data/listAPIData.type";
-import TableAPIData from "../../types/data/tableAPIData.type";
+import ListAPIData from "../../types/listAPIData.type";
+import TableAPIData from "../../types/tableAPIData.type";
 import SelectType from "../../types/selectType.type";
 
 type GameAreaProps = {
