+++ src/components/SelectOptions/type.ts
@@ -1,4 +1,4 @@
-import TableAPIData from "../../types/data/tableAPIData.type";
+import TableAPIData from "../../types/tableAPIData.type";
 import SelectType from "../../types/selectType.type";
 
 type SelectOptionsProps = {
