+++ src/pages/WeaponPage/type.ts
@@ -1,5 +1,5 @@
 import { GameDataType, WeaponData } from "../../__generated__/graphql";
-import TableAPIData from "../../types/data/tableAPIData.type";
+import TableAPIData from "../../types/tableAPIData.type";
 
 type WeaponPageProps = {
   dailyRecordID: string;
