+++ src/types/listAPIData.type.ts
@@ -1,4 +1,4 @@
-import { ConstellationData, TalentData } from "../../__generated__/graphql";
+import { ConstellationData, TalentData } from "../__generated__/graphql";
 
 type ListAPIData = TalentData | ConstellationData;
