+++ src/types/tableAPIData.type.ts
@@ -1,8 +1,4 @@
-import {
-  CharacterData,
-  FoodData,
-  WeaponData,
-} from "../../__generated__/graphql";
+import { CharacterData, FoodData, WeaponData } from "../__generated__/graphql";
 
 type TableAPIData = CharacterData | WeaponData | FoodData;
